Standardize Module Install (#23)
Upgraded to OpenZeppelin contracts v4.8.3 and there is a breaking change
for `ERC721.sol` in that repo as of v4.8.2. This PR fixes the `_batchMint` functions
for `ERC721TL` and `Shatter` to account for this.

Also, updated one test for `ERC1155TL.t.sol` so that there were no
random failures.