Version 2.5.0 (#27)
Added batch mint functionality that does not use ERC-2309 in `ERC721TL`
and `Shatter`.

## ERC721TL
- migrated to using `Transfer` events on the `batchMint` function
- created a `batchMintUltra` function that can be used if the target
platform listens to ERC-2309's `ConsecutiveTransfer` event. This should
not be the default function to use.

## Shatter
- migrated to using `Transfer` events on the `shatter` function
- created a `shatterUltra` function that uses ERC-2309's
`ConsecutiveTransfer` event. This should not be the default function to
use.

Tests have been updated to reflect this and 1024 fuzz runs were
completed with passing results. (minor updates included to the test
suite to ensure passing results).

Updated gas cost for `batchMint` on `ERC721TL`:
| NFTs | Gas Cost |
|------|-----------|
| 5 | 198,000 |
| 10 | 209,000 |
| 100 | 402,000 |
| 200 | 616,000 |
| 500 | 1,259,000 |
| 1000 | 2,331,000 |